+++ scripts/relic_evaluation_crealm.py
@@ -144,6 +144,7 @@ for book_title, book_data in data.items():
         if not load_existing:
             similarities = tf.matmul(all_prefices[ns], tf.transpose(all_suffixes[ns]))
             sorted_scores = tf.argsort(similarities, axis=1, direction='DESCENDING')
+            sorted_score_vals = tf.sort(similarities, axis=1, direction='DESCENDING')
 
         ranks = []
         for qnum, (quote, context) in enumerate(all_quotes_len_dict[ns]):
@@ -190,7 +191,7 @@ for book_title, book_data in data.items():
         )
     print("")
 
-if not load_existing:
+if not load_existing and args.write_to_file:
     with open(f"{args.output_dir}/{args.split}_with_ranks_crealm_left_{args.left_sents}_right_{args.right_sents}.json", "w") as f:
         f.write(json.dumps(data))
